requirement.txt
gspread
oauth2client
selenium
beautifulsoup4
openai==0.28
python-dotenv
pinecone-client==3.1.0